# _archive/api/v1/issue_handler.go
package v1

import (
	"net/http"

	"github.com/despiegk/goweb2/internal/model"
	"github.com/gin-gonic/gin"
)

// @Summary Get all issues
// @Description Get a list of all issues
// @Tags issues
// @Accept json
// @Produce json
// @Success 200 {array} model.Issue
// @Router /issues [get]
func GetIssues(c *gin.Context) {
	// TODO: Implement database integration
	issues := []model.Issue{}
	c.JSON(http.StatusOK, issues)
}

// @Summary Create a new issue
// @Description Create a new issue with the provided information
// @Tags issues
// @Accept json
// @Produce json
// @Param issue body model.Issue true "Issue object"
// @Success 201 {object} model.Issue
// @Router /issues [post]
func CreateIssue(c *gin.Context) {
	var issue model.Issue
	if err := c.ShouldBindJSON(&issue); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement database integration
	c.JSON(http.StatusCreated, issue)
}

// @Summary Get an issue by ID
// @Description Get detailed information about a specific issue
// @Tags issues
// @Accept json
// @Produce json
// @Param id path string true "Issue ID"
// @Success 200 {object} model.Issue
// @Router /issues/{id} [get]
func GetIssue(c *gin.Context) {
	id := c.Param("id")
	// TODO: Implement database integration - fetch issue with id
	_ = id                 // Will be used when implementing database operations
	issue := model.Issue{} // Placeholder for issue with id
	c.JSON(http.StatusOK, issue)
}

// @Summary Update an issue
// @Description Update an existing issue with new information
// @Tags issues
// @Accept json
// @Produce json
// @Param id path string true "Issue ID"
// @Param issue body model.Issue true "Issue object"
// @Success 200 {object} model.Issue
// @Router /issues/{id} [put]
func UpdateIssue(c *gin.Context) {
	id := c.Param("id")
	var issue model.Issue
	if err := c.ShouldBindJSON(&issue); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement database integration - update issue with id
	_ = id // Will be used when implementing database operations
	c.JSON(http.StatusOK, issue)
}

// @Summary Delete an issue
// @Description Delete an issue by ID
// @Tags issues
// @Accept json
// @Produce json
// @Param id path string true "Issue ID"
// @Success 204 "No Content"
// @Router /issues/{id} [delete]
func DeleteIssue(c *gin.Context) {
	id := c.Param("id")
	// TODO: Implement database integration - delete issue with id
	_ = id // Will be used when implementing database operations
	c.Status(http.StatusNoContent)
}

// @Summary Add time entry to an issue
// @Description Add a new time entry to an existing issue
// @Tags issues
// @Accept json
// @Produce json
// @Param id path string true "Issue ID"
// @Param timeEntry body model.TimeEntry true "Time Entry object"
// @Success 201 {object} model.TimeEntry
// @Router /issues/{id}/time [post]
func AddTimeEntry(c *gin.Context) {
	id := c.Param("id")
	var timeEntry model.TimeEntry
	if err := c.ShouldBindJSON(&timeEntry); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement database integration - add time entry to issue with id
	_ = id // Will be used when implementing database operations
	c.JSON(http.StatusCreated, timeEntry)
}